exec & builtins further progress
