+++ incl/builtin.h
@@ -16,7 +16,6 @@
 # include <stdlib.h>
 # include "symtab.h"
 
-/* builtins */
 int		bi_cd(char *address, t_symtab *symtab);
 int		bi_echo(char **args);
 int		bi_env(t_symtab *symtab);
@@ -27,5 +26,11 @@ int		bi_unset(char **args, t_symtab *symtab);
 /* builtin_utils.c */
 int		execute_builtin(char **args, t_symtab *symtab);
 int		is_builtin(char *cmd);
+/* builtin_export_lst.c */
+int		display_export_list(t_symtab *symtab);
+/* builtin_export_var.c */
+int		is_valid_var_name(char *var_name);
+char	*get_var_name(char *arg);
+char	*get_var_value(char *arg);
 
 #endif
