+++ src/executor.c
@@ -6,80 +6,18 @@
 /*   By: dmalacov <dmalacov@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/09/08 14:27:06 by dmalacov      #+#    #+#                 */
-/*   Updated: 2022/09/20 18:32:16 by dmalacov      ########   odam.nl         */
+/*   Updated: 2022/09/26 15:53:50 by dmalacov      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "main.h"
-#include "libft.h"
+#include "minishell.h"
+// #include "libft.h"	is in minishell.h
 #include "executor.h"
-#include "symtab.h"
+// #include "symtab.h"	is in minishell.h
 #include "builtin.h"
 #include <errno.h>
-#include <stdio.h>
-#include <unistd.h>
-
-static size_t	st_count_cmds(t_token_lst *input)
-{
-	size_t	count;
-
-	count = 0;
-	while (input)
-	{
-		if (input->token_type == PIPE)
-			count++;
-		input = input->next;
-	}
-	return (count + 1);
-}
-
-static int	st_check_if_cmd_builtin(t_token_lst *input, t_cmd_tools *tools)
-{
-	if (tools->total_cmds > 1)
-		return (0);
-	while (input && input->token_type != PIPE)
-	{
-		if (input->token_type == WORD)
-		{
-			if (is_builtin(input->content))
-				return (1);
-			else
-				return (0);
-		}
-		else if (input->token_type == GRT_TH || input->token_type == DGRT_TH || \
-		input->token_type == SMLR_TH || input->token_type == DSMLR_TH)
-			input = input->next;
-		input = input->next;
-	}	
-	return (0);
-}
-
-static t_cmd_tools	*st_tools_init(t_token_lst *input, t_symtab *symtab)
-{
-	t_cmd_tools	*tools;
-
-	tools = malloc(sizeof(t_cmd_tools));
-	if (!tools)
-		return (NULL);
-	tools->id = 1;
-	tools->cmd = 1;
-	tools->total_cmds = st_count_cmds(input);
-	tools->input_fd = STDIN_FILENO;
-	tools->output_fd = STDOUT_FILENO;
-	tools->cmd_args = NULL;
-	if (get_paths(symtab, tools) == 1 || get_env_var(symtab, tools) == 1)
-		return (cleanup(tools), NULL);
-	if (check_heredoc(input, tools) == 1 || get_heredoc(tools->heredoc) == 1)
-		return (cleanup(tools), NULL);
-	tools->builtin_only = st_check_if_cmd_builtin(input, tools);
-	tools->process_tokens[WORD] = process_word;
-	tools->process_tokens[GRT_TH] = process_output_redir1;
-	tools->process_tokens[SMLR_TH] = process_input_redir1;
-	tools->process_tokens[DGRT_TH] = process_output_redir2;
-	tools->process_tokens[DSMLR_TH] = process_input_redir2;
-	tools->process_tokens[EMPTY] = process_word;
-	return (tools);
-}
+// #include <stdio.h>	is in minishell.h
+// #include <unistd.h>	is in minishell.h
 
 static t_token_lst	*st_goto_nxt_cmd(t_token_lst *node)
 {
@@ -90,6 +28,24 @@ static t_token_lst	*st_goto_nxt_cmd(t_token_lst *node)
 	return (node);
 }
 
+int	st_prepare_to_exit(t_cmd_tools *tools, t_token_lst *input, \
+t_symtab *symtab, int exit_code)
+{
+	if (exit_code == -1)
+		exit_code = parent_exec_builtin(tools, input, symtab);
+	else if (exit_code == 0)
+		exit_code = wait_for_last_child(tools->id, tools->total_cmds);
+	else if (exit_code == 1)
+	{
+		while (tools->cmd > 1)
+		{
+			wait(NULL);
+			tools->cmd--;
+		}
+	}
+	return (exit_code);
+}
+
 int	executor(t_token_lst *input, t_symtab *symtab)
 {
 	t_token_lst	*node;
@@ -97,7 +53,7 @@ int	executor(t_token_lst *input, t_symtab *symtab)
 	int			pipe_end[2][2];
 	int			exit_code;
 
-	tools = st_tools_init(input, symtab);
+	tools = tools_init(input, symtab);
 	if (!tools)
 		return (1);
 	node = input;
@@ -113,11 +69,7 @@ int	executor(t_token_lst *input, t_symtab *symtab)
 		tools->cmd++;
 		node = st_goto_nxt_cmd(node);
 	}
-	if (exit_code == 0 && tools->id > 0)
-		exit_code = wait_for_last_child(tools->id, tools->total_cmds);
-	if (exit_code == -1)
-		exit_code = parent_exec_builtin(tools, input, symtab);
+	exit_code = st_prepare_to_exit(tools, input, symtab, exit_code);
 	cleanup(tools);
-	// printf("Executor returning exit code %d\n", exit_code);	// delete
 	return (exit_code);
 }
